Finished staff hire dialog, with the exception of the actual hiring button. Tweaked staff wage calculations. Fixed math.random bugs with negative minimums. Fixed x64 compile warning in LFS.

+++ LFS/lfs.c
@@ -82,7 +82,7 @@
 typedef struct dir_data {
 	int  closed;
 #ifdef _WIN32
-	long hFile;
+	intptr_t hFile;
 	char pattern[MAX_DIR_LENGTH+1];
 #else
 	DIR *dir;
